Defaulting behavior for currency field should not be dependent on amount fields
It does not make sense to only return the default when there is an amount set.
When the currency gets set explicitly, and there's no amount, the currency still
gets returned. It seems best that the default value and custom values should
have the same behavior vis a vis what gets returned when no amount is set.